First of dozens no doubt!
First commit & push, though this is also just the placeholder and to maybe give Tara something visual to work with - minimal as it is - because not even knowing what the *basic* form of the pages will be might make it hard to "commit" to a colour/layout.

See what I did there ho! :)